// src/common/types.ts
export type Arbundles = {
  createData: typeof createData;
  DataItem: typeof DataItem;
  deepHash: typeof deepHash;
  stringToBuffer: typeof stringToBuffer;
  getCryptoDriver: typeof getCryptoDriver;
  bundleAndSignData: typeof bundleAndSignData;
};

export type ARxTransaction = {
  sign: () => Promise<Buffer>;
  size: number;
  uploadWithReceipt: (opts?: UploadOptions) => Promise<UploadReceipt>;
  upload(opts: UploadOptions & { getReceiptSignature: true }): Promise<UploadReceipt>;
  upload(opts?: UploadOptions): Promise<UploadResponse>;
  isValid(): Promise<boolean>;
  // fromRaw(rawTransaction: Buffer, ARxInstance: ARx): ARxTransaction;
} & DataItem;
export type ARxTransactonCtor = new (
  data: string | Uint8Array,
  ARx: Pick<ARx, "uploader" | "tokenConfig" | "arbundles">,
  opts?: ARxTransactionCreateOptions,
) => ARxTransaction;

export type Tx = {
  from: string;
  to: string;
  amount: BigNumber;
  blockHeight?: BigNumber;
  pending: boolean;
  confirmed: boolean;
};
export type TokenConfig<Wallet = string | object, Opts = any> = {
  arx: ARx;
  name: string;
  ticker: string;
  minConfirm?: number;
  wallet?: Wallet;
  providerUrl: string;
  isSlow?: boolean;
  opts?: Opts;
};

export type Network = "mainnet" | "devnet" | string;

export type ARxConfig = {
  timeout?: number;
  providerUrl?: string;
  contractAddress?: string;
  tokenOpts?: object;
  headers?: Record<string, string>;
  debug?: boolean;
};

export type Token = {
  isSlow: boolean;
  needsFee: boolean;

  base: [string, number];

  name: string;

  get address(): string | undefined;

  ticker: string;

  arx: ARx;

  getTx(txId: string): Promise<Tx>;

  ownerToAddress(owner: any): string;

  price(): Promise<number>;

  sign(data: Uint8Array): Promise<Uint8Array>;

  getSigner(): Signer;

  verify(pub: any, data: Uint8Array, signature: Uint8Array): Promise<boolean>;

  getCurrentHeight(): Promise<BigNumber>;

  getFee(amount: BigNumber.Value, to?: string, multiplier?: BigNumber.Value): Promise<BigNumber | object>;

  sendTx(data: any): Promise<string | undefined>;

  createTx(amount: BigNumber.Value, to: string, fee?: string | object): Promise<{ txId: string | undefined; tx: any }>;

  getPublicKey(): Promise<string | Buffer> | (string | Buffer);

  ready?(): Promise<void>;

  // createMultiSigTx?(amount: BigNumber.Value, to: string, opts: any, fee?: string): Promise<any>;

  // addSignature?(multiTx: any, opts: any): Promise<any>;

  // submitMultiSigTx?(multiTx: any, opts: any): Promise<any>;
};

export type Manifest = {
  manifest: string;
  version: string;
  paths: Record<string, Record<string, Record<"id", string>>>;
  index?: Record<"path", string>;
};

export type UploadResponse = {
  // The ID of the transaction
  id: string;
  // The Arweave public key of the node that received the transaction
  public: string;
  // The signature of this receipt
  signature: string;
  // the maximum expected Arweave block height for transaction inclusion
  deadlineHeight: number;
  // List of validator signatures
  validatorSignatures: { address: string; signature: string }[];
  // The UNIX (MS precision) timestamp of when the node received the Tx.
  timestamp: number;
  // The receipt version
  version: "1.0.0";
  // Injected verification function (same as Utils/ARx.verifyReceipt)
  verify: () => Promise<boolean>;
};

export type UploadReceipt = /* Required<UploadResponse>; */ UploadResponse;
export type UploadReceiptData = Omit<UploadReceipt, "verify" | "validatorSignatures">;

export type FundResponse = {
  reward: string;
  target: string;
  quantity: string;
  id: string;
};
export type WithdrawalResponse = {
  tx_id: string;
  requested: number;
  fee: number;
  final: number;
};

export type CreateAndUploadOptions = DataItemCreateOptions & { upload?: UploadOptions };

export type UploadOptions = { paidBy?: string };

export enum UploadHeaders {
  PAID_BY = "x-arx-paid-by",
}

// // TS doesn't like string template literals it seems
// export enum manifestType {
//     paths = "arweave/paths"
// }

// export enum manifestVersion {
//     "0.1.0" = "0.1.0"
// }

export type ARxTransactionCreateOptions = DataItemCreateOptions & { dataIsRawTransaction?: boolean };